updated module docs, moved rtd/thermal stuff to thermal module, moved simulation stuff to sim
